Update index.html.erb
I think having a "New Game" button without any additional details is a little confusing on the root page, and probably not necessary. From a UX perspective, the player has a clear place to go if they want to start a new game: The "Start A New Game" button in the navbar.